Add Frame::BlockLogical and Frame::ElementLogical
There are actually two logical frames used in spectre, and we should
be able to distinguish between them.  The BlockLogical frame is used to
define the source frame for the CoordinateMaps used to define how a Block
is embedded in a Domain.  The ElementLogical frame is the frame in which
the numerical algorithms are done.  The two frames are related by a
simple affine map.